pyproject.toml
[tool.poetry]
name = "machma10-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Lukas Stevens <mail@lukas-stevens.de>"]
packages = [
  { include = "machma", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.7"
aiogram = "^2.9.2"
tabulate = "^0.8.7"
sqlalchemy = "^1.3.19"
toml = "^0.10.1"
"databind.core" = "^0.3.0"
"databind.json" = "^0.3.0"
click = "^7.1.2"
deprecated = "^1.2.10"
"nr.proxy" = "1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
